+++ release_notes.md
@@ -12,7 +12,8 @@
 
 ### ✨ New Functionality
 
-- Support service bindings to the [SAP BTP AI Core Service](https://api.sap.com/api/AI_CORE_API) by default in the `ServiceBindingDestinationLoader` API. 
+- Support service bindings to the [SAP BTP AI Core Service](https://api.sap.com/api/AI_CORE_API) by default in the `ServiceBindingDestinationLoader` API.
+- Failed OData v4 Batch requests now return the specific failed request from the exception: `ODataResponseException.getRequest()`.
 
 ### 📈 Improvements